AP-8251: Remove code of shrinker node
This is necessary because otherwise there would be a dependency cycle
between the ensembles and the treeensembles plugin.
As it is slow anyway and not even in the AP currently, this should not
cause any problems.
If we need it again, we can still access it on the release branches
prior to 3.5.

AP-8251 (Add Gradient Boosted Trees and Simple Regression Tree support
to generic PMML Predictor)