AP-18404: Allow garbage collector to collect TreeDataCreator
Not really related to the ticket but still problematic.
The code held onto the TreeDataCreator which contains very large data
structures that aren't used after creation of the TreeData. This can
cause memory issues if there is a lot of data.

AP-18404 (Gradient Boosted Tree error: Failed: Exponential Sum is not
finite)